hotfix: header 검색 버그 수정 및 스타일 수정
